Gacha fixes, guild fixes
Corrects db stuff for shops and gacha being written as if a single person was on the server.

Fixes netcafe points being a sized ack instead of just a simple which was breaking the blacksmith look up despite working on the actual store.

Adds missing timestamp to entrance response and a basic canned USR field response. This fixes both entering guild halls and the quest lists not having all entries they should have.

Adds login boost support.

Writes GR level as LE to the login response so it's displayed correctly in launcher.

A bunch of the guild hall packets should probably also be moved into handlers_guild.go but I didn't get around to it.

Similarly the canned responses are currently set up for a max level guild with all facilities and selective responses will need to be sorted at some point.

As the USR response is just responding with server 1 channel 1 it's probably advisable to remove the noob server from config all together for the moment if you actually want access to a guild hall.